io: use large buffer size
This update uses a larger buffer size in the i/o streams. The default
size is 4096. This cuts down on the number of times we cross the JNI
boundary. As an example running teravalidate on a 10GB data set won't
finish even after 10 or 15 minutes, but with a 2MB buffer it finishes in
2 minutes. Most likely because teravalidate is doing tiny reads.

When we support ByteBuffer interface we should be able to get rid of the
use of the java buffer completely and rely on buffering and caching in
libcephfs.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>